# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GrowthAgent API",
    "version": "1.0.0",
    "description": "Autonomous AI growth system with ONE deterministic gatekeeper. Built for Razorpay AI Buildathon."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development & demo server"
    }
  ],
  "paths": {
...
    "/v1/admin/approvals": {
      "get": {
        "summary": "List Escalations Inbox",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending or resolved escalations."
          }
        }
      }
    },
    "/v1/admin/approvals/{id}/approve": {
      "post": {
        "summary": "Human Approve Escalation",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Settlement resumes with the SAME frozen proposal."
          },
          "409": {
            "description": "Already resolved or rules drifted."
          }
        }
      }
    },
    "/v1/admin/approvals/{id}/reject": {
      "post": {
        "summary": "Human Reject Escalation",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Proposal marked terminal DECLINED."
          },
          "409": {
            "description": "Already resolved."
          }
        }
      }
    },
    "/v1/admin/audit/{txId}/replay": {
      "get": {
        "summary": "Replay Hash-Chained Audit Log",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reconstructed timeline and hash chain validity verdict."
          }
        }
      }
    },
...
    "/v1/demo/chaos": {
      "get": {
        "summary": "List Armed Chaos Flags",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ]
      },
      "put": {
        "summary": "Arm Chaos Flag",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Disarm All Chaos Flags",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "RazorpaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Razorpay-Signature"
      }
    }
  }
}